backend/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.6.0
    # via
    #   -r requirements.in
    #   django
    #   djoser
attrs==22.2.0
    # via
    #   -r requirements.in
    #   pytest
build==0.10.0
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2022.12.7
    # via
    #   -r requirements.in
    #   requests
cffi==1.15.1
    # via
    #   -r requirements.in
    #   cryptography
charset-normalizer==2.0.12
    # via
    #   -r requirements.in
    #   requests
click==8.1.3
    # via
    #   -r requirements.in
    #   pip-tools
coreapi==2.3.3
    # via
    #   -r requirements.in
    #   djoser
coreschema==0.0.4
    # via
    #   -r requirements.in
    #   coreapi
cryptography==41.0.1
    # via
    #   -r requirements.in
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   -r requirements.in
    #   python3-openid
    #   social-auth-core
django==3.2
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
django-filter==23.1
    # via -r requirements.in
django-templated-mail==1.1.1
    # via
    #   -r requirements.in
    #   djoser
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.7.2
    # via
    #   -r requirements.in
    #   djoser
djoser==2.1.0
    # via -r requirements.in
idna==3.4
    # via
    #   -r requirements.in
    #   requests
iniconfig==2.0.0
    # via
    #   -r requirements.in
    #   pytest
isort==5.12.0
    # via -r requirements.in
itypes==1.2.0
    # via
    #   -r requirements.in
    #   coreapi
jinja2==3.1.2
    # via
    #   -r requirements.in
    #   coreschema
markupsafe==2.1.3
    # via
    #   -r requirements.in
    #   jinja2
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   requests-oauthlib
    #   social-auth-core
packaging==23.0
    # via
    #   -r requirements.in
    #   build
    #   pytest
pillow==10.0.0
    # via -r requirements.in
pip-tools==6.14.0
    # via -r requirements.in
pluggy==0.13.1
    # via
    #   -r requirements.in
    #   pytest
psycopg2==2.9.6
    # via -r requirements.in
py==1.11.0
    # via
    #   -r requirements.in
    #   pytest
pycparser==2.21
    # via
    #   -r requirements.in
    #   cffi
pyjwt==2.1.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   social-auth-core
pyproject-hooks==1.0.0
    # via
    #   -r requirements.in
    #   build
pytest==6.2.4
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-pythonpath
pytest-django==4.4.0
    # via -r requirements.in
pytest-pythonpath==0.7.3
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
python3-openid==3.2.0
    # via
    #   -r requirements.in
    #   social-auth-core
pytz==2023.3
    # via
    #   -r requirements.in
    #   django
requests==2.26.0
    # via
    #   -r requirements.in
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via
    #   -r requirements.in
    #   social-auth-core
six==1.16.0
    # via
    #   -r requirements.in
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via
    #   -r requirements.in
    #   djoser
social-auth-core==4.4.2
    # via
    #   -r requirements.in
    #   social-auth-app-django
sqlparse==0.4.3
    # via
    #   -r requirements.in
    #   django
toml==0.10.2
    # via
    #   -r requirements.in
    #   pytest
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   coreapi
urllib3==1.26.15
    # via
    #   -r requirements.in
    #   requests
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
